change package name to lower case.
