+++ .idea/compiler.xml
@@ -6,7 +6,7 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="FileChampion4j" />
+        <module name="filechampion4j" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
