+++ src/main/java/com/blumo/filechampion4j/PluginsHelper.java
@@ -1,4 +1,4 @@
-package com.blumo.FileChampion4j;
+package com.blumo.filechampion4j;
 
 import org.json.*;
 import java.util.HashMap;
