+++ src/test/java/com/blumo/filechampion4j/FileValidatorLargeFilesBench.java
@@ -1,4 +1,5 @@
-package com.blumo.FileChampion4j;
+package com.blumo.filechampion4j;
+
 
 import org.json.JSONObject;
 import org.junit.jupiter.api.AfterAll;
@@ -25,9 +26,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  * Unit test for FileValidator class.
  */
 
-public class FileValidatorBench {
+public class FileValidatorLargeFilesBench {
 
-    public int batchSize = 100;
+    public int batchSize = 20;
     public FileValidator validator;
     public File[] testFiles;
     public ArrayList<byte[]> fileByteArrayList;
@@ -98,7 +99,7 @@ public class FileValidatorBench {
             fileNamesArray = new String[batchSize + warmupRepeats + 1];
 
             for (int i=0; i < batchSize + warmupRepeats + 1; i++ ) {
-                fileByteArrayList.add(generatePdfBytes(300000));
+                fileByteArrayList.add(generatePdfBytes(1000000));
                 fileNamesArray[i] = UUID.randomUUID().toString() + ".pdf";
             }
